♻️ refactor: Separate the list of live products
